aya: rough first cut at userspace ringbuf API
This API closely follows the [libbpf] ringbuf semantics but represents a pretty big
departure from how we handle perf buffers at the moment. In particular, we rely on
a callback-based approach to consuming events.

This code is probably full of bugs and needs to be properly tested before it can be
merged. The API could also probably use some polish, and we may end up deciding that more
consistency with perf buffer is desired.

Currently there is no async support. This will come later.

[libbpf]: https://github.com/libbpf/libbpf/blob/master/src/ringbuf.c